[Feature] to #49886785 support switch Pg/Ora mode Link: https://code.alibaba-inc.com/PolarDB-for-PostgreSQL/polardb_jdbc/codereview/12931547 * [Feature] to #49886785 support switch Pg/Ora mode
Since the introduction of a set of community jar packages
in the 2.0 jdbc may cause the failure of external jar packages,
the community jar packages are no longer introduced;

in order to solve this problem, the pg/ora mode is introduced
on the same set of codes, through the connection header
to distinguish the driver.

* [Feature] to #49886785 support switch Pg/Ora mode

Since the introduction of a set of community jar packages
in the 2.0 jdbc may cause the failure of external jar packages,
the community jar packages are no longer introduced;

in order to solve this problem, the pg/ora mode is introduced
on the same set of codes, through the connection header
to distinguish the driver.

* [Feature] to #49886785 support switch Pg/Ora mode

Since the introduction of a set of community jar packages
in the 2.0 jdbc may cause the failure of external jar packages,
the community jar packages are no longer introduced;

in order to solve this problem, the pg/ora mode is introduced
on the same set of codes, through the connection header
to distinguish the driver.

* [Feature] to #49886785 support switch Pg/Ora mode

Since the introduction of a set of community jar packages
in the 2.0 jdbc may cause the failure of external jar packages,
the community jar packages are no longer introduced;

in order to solve this problem, the pg/ora mode is introduced
on the same set of codes, through the connection header
to distinguish the driver.

* [Feature] to #49886785 support switch Pg/Ora mode

Since the introduction of a set of community jar packages
in the 2.0 jdbc may cause the failure of external jar packages,
the community jar packages are no longer introduced;

in order to solve this problem, the pg/ora mode is introduced
on the same set of codes, through the connection header
to distinguish the driver.

* [Feature] to #49886785 support switch Pg/Ora mode

Since the introduction of a set of community jar packages
in the 2.0 jdbc may cause the failure of external jar packages,
the community jar packages are no longer introduced;

in order to solve this problem, the pg/ora mode is introduced
on the same set of codes, through the connection header
to distinguish the driver.

* [Feature] to #49886785 support switch Pg/Ora mode

Since the introduction of a set of community jar packages
in the 2.0 jdbc may cause the failure of external jar packages,
the community jar packages are no longer introduced;

in order to solve this problem, the pg/ora mode is introduced
on the same set of codes, through the connection header
to distinguish the driver.

* [Feature] to #49886785 support switch Pg/Ora mode

Since the introduction of a set of community jar packages
in the 2.0 jdbc may cause the failure of external jar packages,
the community jar packages are no longer introduced;

in order to solve this problem, the pg/ora mode is introduced
on the same set of codes, through the connection header
to distinguish the driver.